+++ src/cli.ts
@@ -25,6 +25,7 @@ import { createTicketsCommand } from "./commands/tickets.js";
 import { createContractsCommand } from "./commands/contracts.js";
 import { createShipmentsCommand } from "./commands/shipments.js";
 import { createReceptionsCommand } from "./commands/receptions.js";
+import { createInterventionsCommand } from "./commands/interventions.js";
 import { createUpgradeCommand } from "./commands/upgrade.js";
 import {
   ensureFreshCacheOnColdStart,
@@ -81,6 +82,7 @@ program.addCommand(createTicketsCommand());
 program.addCommand(createContractsCommand());
 program.addCommand(createShipmentsCommand());
 program.addCommand(createReceptionsCommand());
+program.addCommand(createInterventionsCommand());
 program.addCommand(createUpgradeCommand());
 
 configureHelp(program);